Revert "streamingccl: move testutils into their own package"
This reverts commit 606aa9ae3cc75a7b278f5f7e7346e2af571fd026.
