+++ pkg/BUILD.bazel
@@ -82,10 +82,10 @@ ALL_TESTS = [
     "//pkg/ccl/sqlproxyccl:sqlproxyccl_test",
     "//pkg/ccl/storageccl/engineccl:engineccl_test",
     "//pkg/ccl/storageccl:storageccl_test",
-    "//pkg/ccl/streamingccl/replicationutils:replicationutils_test",
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
     "//pkg/ccl/streamingccl/streamingest:streamingest_test",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
+    "//pkg/ccl/streamingccl:streamingccl_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
     "//pkg/ccl/testccl/sqlccl:sqlccl_test",
@@ -801,17 +801,15 @@ GO_TARGETS = [
     "//pkg/ccl/storageccl/engineccl:engineccl_test",
     "//pkg/ccl/storageccl:storageccl",
     "//pkg/ccl/storageccl:storageccl_test",
-    "//pkg/ccl/streamingccl/replicationtestutils:replicationtestutils",
-    "//pkg/ccl/streamingccl/replicationtestutils:streamingtest",
-    "//pkg/ccl/streamingccl/replicationutils:replicationutils",
-    "//pkg/ccl/streamingccl/replicationutils:replicationutils_test",
     "//pkg/ccl/streamingccl/streamclient:streamclient",
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
     "//pkg/ccl/streamingccl/streamingest:streamingest",
     "//pkg/ccl/streamingccl/streamingest:streamingest_test",
+    "//pkg/ccl/streamingccl/streamingtest:streamingtest",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer",
     "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
     "//pkg/ccl/streamingccl:streamingccl",
+    "//pkg/ccl/streamingccl:streamingccl_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
     "//pkg/ccl/testccl/sqlccl:sqlccl_test",
@@ -2316,10 +2314,9 @@ GET_X_DATA_TARGETS = [
     "//pkg/ccl/storageccl/engineccl:get_x_data",
     "//pkg/ccl/storageccl/engineccl/enginepbccl:get_x_data",
     "//pkg/ccl/streamingccl:get_x_data",
-    "//pkg/ccl/streamingccl/replicationtestutils:get_x_data",
-    "//pkg/ccl/streamingccl/replicationutils:get_x_data",
     "//pkg/ccl/streamingccl/streamclient:get_x_data",
     "//pkg/ccl/streamingccl/streamingest:get_x_data",
+    "//pkg/ccl/streamingccl/streamingtest:get_x_data",
     "//pkg/ccl/streamingccl/streamproducer:get_x_data",
     "//pkg/ccl/telemetryccl:get_x_data",
     "//pkg/ccl/testccl/authccl:get_x_data",
