--- pkg/ccl/streamingccl/replicationtestutils/BUILD.bazel
@@ -1,76 +1,0 @@
-load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "streamingtest",
-    srcs = [
-        "encoding.go",
-        "replication_helpers.go",
-    ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamingtest",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/base",
-        "//pkg/ccl/changefeedccl/changefeedbase",
-        "//pkg/ccl/streamingccl",
-        "//pkg/jobs/jobspb",
-        "//pkg/keys",
-        "//pkg/repstream/streampb",
-        "//pkg/roachpb",
-        "//pkg/security/username",
-        "//pkg/sql/catalog",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/desctestutils",
-        "//pkg/sql/rowenc",
-        "//pkg/sql/sem/tree",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/util/hlc",
-        "//pkg/util/protoutil",
-        "//pkg/util/syncutil",
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_stretchr_testify//require",
-    ],
-)
-
-go_library(
-    name = "replicationtestutils",
-    srcs = [
-        "encoding.go",
-        "replication_helpers.go",
-        "testutils.go",
-    ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/replicationtestutils",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/base",
-        "//pkg/ccl/changefeedccl/changefeedbase",
-        "//pkg/ccl/streamingccl",
-        "//pkg/jobs",
-        "//pkg/jobs/jobspb",
-        "//pkg/keys",
-        "//pkg/repstream/streampb",
-        "//pkg/roachpb",
-        "//pkg/security/username",
-        "//pkg/server/serverpb",
-        "//pkg/sql",
-        "//pkg/sql/catalog",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/desctestutils",
-        "//pkg/sql/execinfra",
-        "//pkg/sql/rowenc",
-        "//pkg/sql/sem/tree",
-        "//pkg/testutils",
-        "//pkg/testutils/jobutils",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/testutils/testcluster",
-        "//pkg/util/hlc",
-        "//pkg/util/protoutil",
-        "//pkg/util/syncutil",
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_stretchr_testify//require",
-    ],
-)
-
-get_x_data(name = "get_x_data")
