+++ pkg/ccl/streamingccl/streamclient/BUILD.bazel
@@ -54,7 +54,8 @@ go_test(
         "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/ccl/storageccl",
         "//pkg/ccl/streamingccl",
-        "//pkg/ccl/streamingccl/replicationtestutils",
+        "//pkg/ccl/streamingccl/streamingtest",
+        "//pkg/ccl/streamingccl/streamproducer",
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/repstream/streampb",
