+++ pkg/ccl/streamingccl/streamingest/BUILD.bazel
@@ -12,14 +12,15 @@ go_library(
         "stream_ingestion_planning.go",
         "stream_ingestion_processor.go",
         "stream_ingestion_processor_planning.go",
+        "stream_ingestion_test_utils.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamingest",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/base",
         "//pkg/ccl/backupccl",
+        "//pkg/ccl/changefeedccl/cdctest",
         "//pkg/ccl/streamingccl",
-        "//pkg/ccl/streamingccl/replicationutils",
         "//pkg/ccl/streamingccl/streamclient",
         "//pkg/ccl/utilccl",
         "//pkg/jobs",
@@ -53,6 +54,7 @@ go_library(
         "//pkg/sql/types",
         "//pkg/storage",
         "//pkg/storage/enginepb",
+        "//pkg/testutils/sqlutils",
         "//pkg/util/ctxgroup",
         "//pkg/util/hlc",
         "//pkg/util/log",
@@ -67,6 +69,7 @@ go_library(
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_cockroachdb_redact//:redact",
+        "@com_github_stretchr_testify//require",
     ],
 )
 
@@ -76,11 +79,11 @@ go_test(
     srcs = [
         "alter_replication_job_test.go",
         "main_test.go",
-        "replication_random_client_test.go",
-        "replication_stream_e2e_test.go",
         "stream_ingestion_frontier_processor_test.go",
         "stream_ingestion_job_test.go",
         "stream_ingestion_processor_test.go",
+        "stream_ingestion_test.go",
+        "stream_replication_e2e_test.go",
     ],
     args = ["-test.timeout=895s"],
     embed = [":streamingest"],
@@ -89,14 +92,12 @@ go_test(
         "//pkg/ccl",
         "//pkg/ccl/backupccl",
         "//pkg/ccl/changefeedccl",
-        "//pkg/ccl/changefeedccl/cdctest",
         "//pkg/ccl/changefeedccl/changefeedbase",
         "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/ccl/storageccl",
         "//pkg/ccl/streamingccl",
-        "//pkg/ccl/streamingccl/replicationtestutils",
-        "//pkg/ccl/streamingccl/replicationutils",
         "//pkg/ccl/streamingccl/streamclient",
+        "//pkg/ccl/streamingccl/streamingtest",
         "//pkg/ccl/streamingccl/streamproducer",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/jobs",
@@ -111,6 +112,7 @@ go_test(
         "//pkg/security/securitytest",
         "//pkg/security/username",
         "//pkg/server",
+        "//pkg/server/serverpb",
         "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/catalog/desctestutils",
