+++ pkg/ccl/streamingccl/streamingest/alter_replication_job_test.go
@@ -13,7 +13,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/replicationtestutils"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/testutils/jobutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
@@ -26,57 +25,57 @@ func TestAlterTenantPauseResume(t *testing.T) {
 	defer log.Scope(t).Close(t)
 
 	ctx := context.Background()
-	args := replicationtestutils.DefaultTenantStreamingClustersArgs
+	args := defaultTenantStreamingClustersArgs
 
-	c, cleanup := replicationtestutils.CreateTenantStreamingClusters(ctx, t, args)
+	c, cleanup := createTenantStreamingClusters(ctx, t, args)
 	defer cleanup()
-	producerJobID, ingestionJobID := c.StartStreamReplication()
+	producerJobID, ingestionJobID := c.startStreamReplication()
 
-	jobutils.WaitForJobToRun(t, c.SrcSysSQL, jobspb.JobID(producerJobID))
-	jobutils.WaitForJobToRun(t, c.DestSysSQL, jobspb.JobID(ingestionJobID))
+	jobutils.WaitForJobToRun(c.t, c.srcSysSQL, jobspb.JobID(producerJobID))
+	jobutils.WaitForJobToRun(c.t, c.destSysSQL, jobspb.JobID(ingestionJobID))
 
-	c.WaitUntilHighWatermark(c.SrcCluster.Server(0).Clock().Now(), jobspb.JobID(ingestionJobID))
+	c.waitUntilHighWatermark(c.srcCluster.Server(0).Clock().Now(), jobspb.JobID(ingestionJobID))
 
 	// Pause the replication job.
-	c.DestSysSQL.Exec(t, `ALTER TENANT $1 PAUSE REPLICATION`, args.DestTenantName)
-	jobutils.WaitForJobToPause(c.T, c.DestSysSQL, jobspb.JobID(ingestionJobID))
+	c.destSysSQL.Exec(t, `ALTER TENANT $1 PAUSE REPLICATION`, args.destTenantName)
+	jobutils.WaitForJobToPause(c.t, c.destSysSQL, jobspb.JobID(ingestionJobID))
 
 	// Unpause the replication job.
-	c.DestSysSQL.Exec(t, `ALTER TENANT $1 RESUME REPLICATION`, args.DestTenantName)
-	jobutils.WaitForJobToRun(c.T, c.DestSysSQL, jobspb.JobID(ingestionJobID))
+	c.destSysSQL.Exec(t, `ALTER TENANT $1 RESUME REPLICATION`, args.destTenantName)
+	jobutils.WaitForJobToRun(c.t, c.destSysSQL, jobspb.JobID(ingestionJobID))
 
-	c.WaitUntilHighWatermark(c.SrcCluster.Server(0).Clock().Now(), jobspb.JobID(ingestionJobID))
+	c.waitUntilHighWatermark(c.srcCluster.Server(0).Clock().Now(), jobspb.JobID(ingestionJobID))
 	var cutoverTime time.Time
-	c.DestSysSQL.QueryRow(t, "SELECT clock_timestamp()").Scan(&cutoverTime)
+	c.destSysSQL.QueryRow(t, "SELECT clock_timestamp()").Scan(&cutoverTime)
 
-	c.DestSysSQL.Exec(c.T, `ALTER TENANT $1 COMPLETE REPLICATION TO SYSTEM TIME $2::string`, args.DestTenantName, cutoverTime)
-	jobutils.WaitForJobToSucceed(c.T, c.DestSysSQL, jobspb.JobID(ingestionJobID))
-	cleanupTenant := c.CreateDestTenantSQL(ctx)
+	c.destSysSQL.Exec(c.t, `ALTER TENANT $1 COMPLETE REPLICATION TO SYSTEM TIME $2::string`, args.destTenantName, cutoverTime)
+	jobutils.WaitForJobToSucceed(c.t, c.destSysSQL, jobspb.JobID(ingestionJobID))
+	cleanupTenant := c.createDestTenantSQL(ctx)
 	defer func() {
 		require.NoError(t, cleanupTenant())
 	}()
 
 	t.Run("pause-nonexistant-tenant", func(t *testing.T) {
-		c.DestSysSQL.ExpectErr(t, "tenant \"nonexistent\" does not exist", `ALTER TENANT $1 PAUSE REPLICATION`, "nonexistent")
+		c.destSysSQL.ExpectErr(t, "tenant \"nonexistent\" does not exist", `ALTER TENANT $1 PAUSE REPLICATION`, "nonexistent")
 	})
 
 	t.Run("pause-resume-tenant-with-no-replication", func(t *testing.T) {
-		c.DestSysSQL.Exec(t, `CREATE TENANT noreplication`)
-		c.DestSysSQL.ExpectErr(t, "tenant \"noreplication\" does not have an active replication job",
+		c.destSysSQL.Exec(t, `CREATE TENANT noreplication`)
+		c.destSysSQL.ExpectErr(t, "tenant \"noreplication\" does not have an active replication job",
 			`ALTER TENANT $1 PAUSE REPLICATION`, "noreplication")
-		c.DestSysSQL.ExpectErr(t, "tenant \"noreplication\" does not have an active replication job",
+		c.destSysSQL.ExpectErr(t, "tenant \"noreplication\" does not have an active replication job",
 			`ALTER TENANT $1 RESUME REPLICATION`, "noreplication")
 	})
 
 	t.Run("pause-resume-in-readonly-txn", func(t *testing.T) {
-		c.DestSysSQL.Exec(t, `set default_transaction_read_only = on;`)
-		c.DestSysSQL.ExpectErr(t, "cannot execute ALTER TENANT REPLICATION in a read-only transaction", `ALTER TENANT $1 PAUSE REPLICATION`, "foo")
-		c.DestSysSQL.ExpectErr(t, "cannot execute ALTER TENANT REPLICATION in a read-only transaction", `ALTER TENANT $1 RESUME REPLICATION`, "foo")
-		c.DestSysSQL.Exec(t, `set default_transaction_read_only = off;`)
+		c.destSysSQL.Exec(t, `set default_transaction_read_only = on;`)
+		c.destSysSQL.ExpectErr(t, "cannot execute ALTER TENANT REPLICATION in a read-only transaction", `ALTER TENANT $1 PAUSE REPLICATION`, "foo")
+		c.destSysSQL.ExpectErr(t, "cannot execute ALTER TENANT REPLICATION in a read-only transaction", `ALTER TENANT $1 RESUME REPLICATION`, "foo")
+		c.destSysSQL.Exec(t, `set default_transaction_read_only = off;`)
 	})
 
 	t.Run("pause-resume-as-non-system-tenant", func(t *testing.T) {
-		c.DestTenantSQL.ExpectErr(t, "only the system tenant can alter tenant", `ALTER TENANT $1 PAUSE REPLICATION`, "foo")
-		c.DestTenantSQL.ExpectErr(t, "only the system tenant can alter tenant", `ALTER TENANT $1 RESUME REPLICATION`, "foo")
+		c.destTenantSQL.ExpectErr(t, "only the system tenant can alter tenant", `ALTER TENANT $1 PAUSE REPLICATION`, "foo")
+		c.destTenantSQL.ExpectErr(t, "only the system tenant can alter tenant", `ALTER TENANT $1 RESUME REPLICATION`, "foo")
 	})
 }
