+++ pkg/ccl/streamingccl/streamingtest/BUILD.bazel
@@ -1,0 +1,36 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "streamingtest",
+    srcs = [
+        "encoding.go",
+        "replication_helpers.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamingtest",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/base",
+        "//pkg/ccl/changefeedccl/changefeedbase",
+        "//pkg/ccl/streamingccl",
+        "//pkg/jobs/jobspb",
+        "//pkg/keys",
+        "//pkg/repstream/streampb",
+        "//pkg/roachpb",
+        "//pkg/security/username",
+        "//pkg/sql/catalog",
+        "//pkg/sql/catalog/descpb",
+        "//pkg/sql/catalog/desctestutils",
+        "//pkg/sql/rowenc",
+        "//pkg/sql/sem/tree",
+        "//pkg/testutils/serverutils",
+        "//pkg/testutils/sqlutils",
+        "//pkg/util/hlc",
+        "//pkg/util/protoutil",
+        "//pkg/util/syncutil",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
+    ],
+)
+
+get_x_data(name = "get_x_data")
