+++ pkg/ccl/streamingccl/utils_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package replicationutils
+package streamingccl
 
 import (
 	"sort"
